+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.0] - 2017-08-24
+
+### Changed
+* removed custom module reader module and use parseArticles Hook instead
+* disqus not added to article template
+
 ## [1.0.0] - 2017-08-23
 
 * inital release